* Adjust project structure
