+++ .scalafmt.conf
@@ -1,5 +1,3 @@
-
-
 danglingParentheses        = true
 indentOperator             = spray
 maxColumn                  = 100
