+++ Mnist/src/main/resources/application.conf
@@ -1,0 +1,11 @@
+mnist {
+  deeplearning4j {
+    seed = 42
+    numInputs = 784
+    numHidden = 512
+    numOutputs = 10
+    learningRate = 0.01
+    batchSize = 128
+    numEpochs = 10
+  }
+}
