+++ Mnist/src/main/scala/org/mo39/fmbh/mnist/Mnist.scala
@@ -1,0 +1,3 @@
+package org.mo39.fmbh.mnist
+
+object Mnist
