+++ project/plugins.sbt
@@ -1,0 +1,2 @@
+addSbtPlugin("com.typesafe.play" % "sbt-plugin" % "2.6.20")
+addSbtPlugin("ch.epfl.scala" % "sbt-scalafix" % "0.9.1")
